\name{CardinalIO-package}
\alias{CardinalIO-package}
\docType{package}

\alias{CardinalIO}

\title{CardinalIO package}

\description{
Read and write mass spectrometry imaging files
}

\details{
\code{CardinalIO} provides fast and efficient reading and writing of mass spectrometry imaging data files. It supports imzML and Analyze 7.5 formats, and provides ontologies for mass spectrometry imaging.

See \code{vignette("CardinalIO-guide")} for an introduction to the standard imzML format and how to use \code{\link{parseImzML}} and \code{\link{writeImzML}} to parse and write imzML files.

For a complete list of functions, use \code{library(help = "CardinalIO")}.
}

\author{
Kylie A. Bemis
}

\keyword{package}
